Kathy's Sytling Update
Basic changes:
• Logo change
• Footer change including 2 new images
• Removed background image
• Created an images folder and put my images there.  Did not muck with
links to the images you might be using.

NOTE: There are a bunch of ‘artifacts’ in the style that were carried
over when we did the original as a quick-and-dirty to get the GLOS look
and feel.  Since the new GLOS look is much simpler those just aren’t
needed anywhere.